requirements.txt
torch
numpy
pandas
scikit-learn
matplotlib
seaborn
librosa
soundfile
jams
pyyaml
pytest
tensorboard
ipython
torch-summary
tensorboard
itertools